Aggiunta funzione riempi_coda in masterworker, per esplorare i file e directory inseriti come argomenti, rispettando le attese di inserimento con tdelay, che è stato cambiato da tipo int a tipo long per poterlo usare con la funzione sleepTime

+++ header/collector.h
@@ -26,4 +26,4 @@ void printlist();
 void mask_signals_collector();
 int add_res(long sum, char *path);
 void free_res(result_t *res);
-void collector_main(int tdelay);
+void collector_main(long tdelay);
